// src/models/public/SubprojectHistories.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { SubprojectsSubprojectId } from './Subprojects';
import type { AccountsAccountId } from './Accounts';
import type { ProjectsProjectId } from './Projects';

/** Identifier type for public.subproject_histories */
export type SubprojectHistoriesSubprojectHistoryId = string & { __brand: 'public.subproject_histories' };

/**
 * Represents the table public.subproject_histories
 * History of subprojects. Used to analyze subproject development over the years.
 */
export default interface SubprojectHistories {
  subproject_history_id: SubprojectHistoriesSubprojectHistoryId;

  subproject_id: SubprojectsSubprojectId | null;

  /** One history entry per year per subproject */
  year: number | null;

  /** redundant account_id enhances data safety */
  account_id: AccountsAccountId | null;

  project_id: ProjectsProjectId | null;

  name: string | null;

  start_year: number | null;

  end_year: number | null;

  data: unknown | null;

  label: string | null;

  created_at: Date;

  updated_at: Date;

  updated_by: string | null;
}

/**
 * Represents the initializer for the table public.subproject_histories
 * History of subprojects. Used to analyze subproject development over the years.
 */
export interface SubprojectHistoriesInitializer {
  /** Default value: uuid_generate_v7() */
  subproject_history_id?: SubprojectHistoriesSubprojectHistoryId;

  subproject_id?: SubprojectsSubprojectId | null;

  /** One history entry per year per subproject */
  year?: number | null;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  project_id?: ProjectsProjectId | null;

  name?: string | null;

  start_year?: number | null;

  end_year?: number | null;

  data?: unknown | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;

  updated_by?: string | null;
}

/**
 * Represents the mutator for the table public.subproject_histories
 * History of subprojects. Used to analyze subproject development over the years.
 */
export interface SubprojectHistoriesMutator {
  subproject_history_id?: SubprojectHistoriesSubprojectHistoryId;

  subproject_id?: SubprojectsSubprojectId | null;

  /** One history entry per year per subproject */
  year?: number | null;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  project_id?: ProjectsProjectId | null;

  name?: string | null;

  start_year?: number | null;

  end_year?: number | null;

  data?: unknown | null;

  created_at?: Date;

  updated_at?: Date;

  updated_by?: string | null;
}